So here is where I will mention changes I made to the doc.
I'd like it to be super long to see how it will appear in Git!

So here is me extending the line

& extending

for no reason whatsoever.... etc.

Below is an invisible line because I'm going to hash it:
It said "can you see me?" if I cannot, then the hashtag makes it invisible.

But does it make it invisible to the machine too?

I do not know....